:ok_hand: move tests into waitPageActivitiesCompletion into trackPagePageActivity module
