+++ packages/rum/test/trackPageActivities.spec.ts
@@ -1,6 +1,42 @@
-import { RequestCompleteEvent } from '@datadog/browser-core'
+import { noop, Observable, RequestCompleteEvent } from '@datadog/browser-core'
 import { LifeCycle, LifeCycleEventType } from '../src/lifeCycle'
-import { PageActivityEvent, trackPageActivities } from '../src/trackPageActivities'
+import {
+  PAGE_ACTIVITY_END_DELAY,
+  PAGE_ACTIVITY_MAX_DURATION,
+  PAGE_ACTIVITY_VALIDATION_DELAY,
+  PageActivityEvent,
+  trackPageActivities,
+  waitPageActivitiesCompletion,
+} from '../src/trackPageActivities'
+
+// Used to wait some time after the creation of a user action
+const BEFORE_PAGE_ACTIVITY_VALIDATION_DELAY = PAGE_ACTIVITY_VALIDATION_DELAY * 0.8
+// Used to wait some time before the (potential) end of a user action
+const BEFORE_PAGE_ACTIVITY_END_DELAY = PAGE_ACTIVITY_END_DELAY * 0.8
+// A long delay used to wait after any user action is finished.
+const EXPIRE_DELAY = PAGE_ACTIVITY_MAX_DURATION * 10
+
+function mockClock() {
+  beforeEach(() => {
+    jasmine.clock().install()
+    jasmine.clock().mockDate()
+    spyOn(performance, 'now').and.callFake(() => Date.now())
+  })
+
+  afterEach(() => {
+    jasmine.clock().uninstall()
+  })
+
+  return {
+    tick(ms: number) {
+      jasmine.clock().tick(ms)
+    },
+    expire() {
+      // Make sure no user action is still pending
+      jasmine.clock().tick(EXPIRE_DELAY)
+    },
+  }
+}
 
 function eventsCollector<T>() {
   const events: T[] = []
@@ -103,6 +139,117 @@ describe('trackPagePageActivities', () => {
       lifeCycle.notify(LifeCycleEventType.REQUEST_COMPLETED, makeFakeRequestCompleteEvent(11))
       lifeCycle.notify(LifeCycleEventType.REQUEST_COMPLETED, makeFakeRequestCompleteEvent(10))
       expect(events).toEqual([{ isBusy: true }, { isBusy: true }, { isBusy: true }, { isBusy: false }])
+    })
+  })
+})
+
+describe('waitPageActivitiesCompletion', () => {
+  const clock = mockClock()
+
+  it('should not collect an event that is not followed by page activity', (done) => {
+    waitPageActivitiesCompletion(new Observable(), noop, (hadActivity, endTime) => {
+      expect(hadActivity).toBeFalse()
+      expect(endTime).toBeFalsy()
+      done()
+    })
+
+    clock.expire()
+  })
+
+  it('should collect an event that is followed by page activity', (done) => {
+    const activityObservable = new Observable<PageActivityEvent>()
+
+    const startTime = performance.now()
+    waitPageActivitiesCompletion(activityObservable, noop, (hadActivity, endTime) => {
+      expect(hadActivity).toBeTrue()
+      expect(endTime).toEqual(startTime + BEFORE_PAGE_ACTIVITY_VALIDATION_DELAY)
+      done()
+    })
+
+    clock.tick(BEFORE_PAGE_ACTIVITY_VALIDATION_DELAY)
+    activityObservable.notify({ isBusy: false })
+
+    clock.expire()
+  })
+
+  describe('extend with activities', () => {
+    it('is extended while there is page activities', (done) => {
+      const activityObservable = new Observable<PageActivityEvent>()
+      const startTime = performance.now()
+
+      // Extend the user action but stops before PAGE_ACTIVITY_MAX_DURATION
+      const extendCount = Math.floor(PAGE_ACTIVITY_MAX_DURATION / BEFORE_PAGE_ACTIVITY_END_DELAY - 1)
+
+      waitPageActivitiesCompletion(activityObservable, noop, (hadActivity, endTime) => {
+        expect(hadActivity).toBeTrue()
+        expect(endTime).toBe(startTime + (extendCount + 1) * BEFORE_PAGE_ACTIVITY_END_DELAY)
+        done()
+      })
+
+      for (let i = 0; i < extendCount; i += 1) {
+        clock.tick(BEFORE_PAGE_ACTIVITY_END_DELAY)
+        activityObservable.notify({ isBusy: false })
+      }
+
+      clock.expire()
+    })
+
+    it('expires after a limit', (done) => {
+      const activityObservable = new Observable<PageActivityEvent>()
+      let stop = false
+      const startTime = performance.now()
+
+      // Extend the user action until it's more than PAGE_ACTIVITY_MAX_DURATION
+      const extendCount = Math.ceil(PAGE_ACTIVITY_MAX_DURATION / BEFORE_PAGE_ACTIVITY_END_DELAY + 1)
+
+      waitPageActivitiesCompletion(activityObservable, noop, (hadActivity, endTime) => {
+        expect(hadActivity).toBeTrue()
+        expect(endTime).toBe(startTime + PAGE_ACTIVITY_MAX_DURATION)
+        stop = true
+        done()
+      })
+
+      for (let i = 0; i < extendCount && !stop; i += 1) {
+        clock.tick(BEFORE_PAGE_ACTIVITY_END_DELAY)
+        activityObservable.notify({ isBusy: false })
+      }
+
+      clock.expire()
+    })
+  })
+
+  describe('busy activities', () => {
+    it('is extended while the page is busy', (done) => {
+      const activityObservable = new Observable<PageActivityEvent>()
+      const startTime = performance.now()
+      waitPageActivitiesCompletion(activityObservable, noop, (hadActivity, endTime) => {
+        expect(hadActivity).toBeTrue()
+        expect(endTime).toBe(startTime + BEFORE_PAGE_ACTIVITY_VALIDATION_DELAY + PAGE_ACTIVITY_END_DELAY * 2)
+        done()
+      })
+
+      clock.tick(BEFORE_PAGE_ACTIVITY_VALIDATION_DELAY)
+      activityObservable.notify({ isBusy: true })
+
+      clock.tick(PAGE_ACTIVITY_END_DELAY * 2)
+      activityObservable.notify({ isBusy: false })
+
+      clock.expire()
+    })
+
+    it('expires is the page is busy for too long', (done) => {
+      const activityObservable = new Observable<PageActivityEvent>()
+      const startTime = performance.now()
+      waitPageActivitiesCompletion(activityObservable, noop, (hadActivity, endTime) => {
+        expect(hadActivity).toBeTrue()
+        expect(endTime).toBe(startTime + PAGE_ACTIVITY_MAX_DURATION)
+        done()
+      })
+
+      clock.tick(BEFORE_PAGE_ACTIVITY_VALIDATION_DELAY)
+      activityObservable.notify({ isBusy: true })
+
+      clock.expire()
     })
   })
 })
